+++ mypyc/genopsmain.py
@@ -34,7 +34,7 @@ from mypyc.prebuildvisitor import PreBuildVisitor
 from mypyc.irbuild.vtable import compute_vtable
 from mypyc.irbuild.prepare import build_type_map
 from mypyc.genops import IRBuilder
-from mypyc.genopsvisitor import IRBuilderVisitor
+from mypyc.irbuild.visitor import IRBuilderVisitor
 from mypyc.irbuild.mapper import Mapper
